requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.2.0
pyarrow==15.0.0
azure-storage-blob==12.19.0
python-dotenv==1.0.0
